Use the correct method to grab seconds

+++ app/models/accounts/linkedoauth2token/LinkedOAuth2Token.scala
@@ -26,7 +26,7 @@ case class LinkedOAuth2Token(
   def expiresIn: Option[Int] = maybeExpirationTime.map { expirationTime =>
     val now = OffsetDateTime.now
     if (expirationTime.isAfter(now)) {
-      expirationTime.toEpochSecond - now.toEpochSecond
+      (expirationTime.toEpochSecond - now.toEpochSecond).toInt // OAuth2Info library class uses Int instead of Long
     } else {
       0
     }
